플레이어 데이터 간소화
렉이 조금 사라짐

+++ InversusClient/InversusClient/IVSSCGame.h
@@ -1,8 +1,8 @@
 #pragma once
 
 //게임 세팅
-void SetGame(HWND hWnd, RECT* rectView, RECT* tect, double* dx, double* dy, 
-	int block[][20], double bullet[][4], int* ecolor, double* cx, double* cy,
-	double* seta, int* reload, double effect[][17], double reffect[][4], double enemy[][5], 
+void SetGame(HWND hWnd, RECT* rectView, RECT* tect, float* dx, float* dy, 
+	int block[][20], float bullet[][4], int* ecolor, float* cx, float* cy,
+	float* seta, int* reload, float effect[][17], float reffect[][4], float enemy[][5], 
 	int* ecount, int* etime, BOOL* death, int* dcount, int* life, int* score, int* combo, 
-	int* ten, int* gametime, int level, double sgun[][3], int* scount, BOOL multi);
+	int* ten, int* gametime, int level, float sgun[][3], int* scount, BOOL multi);
